Tests: Overall improvements

+++ tests/integration/tasks/edit.test.js
@@ -4,9 +4,14 @@ import app from '~/app';
 import database from '~/database';
 import { Task } from '~/models';
 import {
-  registerMockAccount,
-  registerMockTask,
+  registerAccount,
+  registerTask,
+  withAuth,
 } from '~tests/utils/integration';
+
+function editTask(taskId) {
+  return withAuth(request(app).put(`/tasks/${taskId}`));
+}
 
 beforeAll(database.connect);
 afterAll(database.disconnect);
@@ -17,33 +22,27 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
     priority: 'low',
   };
 
-  const task = {};
   const account = {};
+  const task = {};
 
   beforeAll(async () => {
-    Object.assign(
-      account,
-      await registerMockAccount({ email: 'edit.tasks@example.com' }),
-    );
+    const registeredAccount = await registerAccount({
+      email: 'edit.tasks@example.com',
+    });
+    Object.assign(account, registeredAccount);
   });
 
   beforeEach(async () => {
-    await Task.deleteMany({});
-    Object.assign(task, await registerMockTask(account));
+    await Task.deleteMany({ owner: account.id });
+
+    const registeredTask = await registerTask(account);
+    Object.assign(task, registeredTask);
   });
 
-  function editTaskRequest(taskId, accessToken) {
-    const ongoingRequest = request(app).put(`/tasks/${taskId}`);
-
-    return accessToken
-      ? ongoingRequest.set('Authorization', `Bearer ${accessToken}`)
-      : ongoingRequest;
-  }
-
   it('should support editing an existing task', async () => {
-    const response = await editTaskRequest(task.id, account.accessToken).send(
-      editFixture,
-    );
+    const response = await editTask(task.id)
+      .auth(account.accessToken)
+      .send(editFixture);
 
     expect(response.status).toBe(204);
 
@@ -57,7 +56,7 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
   });
 
   it('should support editing the name of a task without affecting other fields', async () => {
-    const response = await editTaskRequest(task.id, account.accessToken).send({
+    const response = await editTask(task.id).auth(account.accessToken).send({
       name: editFixture.name,
     });
 
@@ -73,7 +72,7 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
   });
 
   it('should support editing the priority of a task without affecting other fields', async () => {
-    const response = await editTaskRequest(task.id, account.accessToken).send({
+    const response = await editTask(task.id).auth(account.accessToken).send({
       priority: editFixture.priority,
     });
 
@@ -91,23 +90,22 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
   it('should not edit a non-existing task', async () => {
     await Task.findByIdAndRemove(task.id);
 
-    const response = await editTaskRequest(task.id, account.accessToken).send(
-      editFixture,
-    );
+    const response = await editTask(task.id)
+      .auth(account.accessToken)
+      .send(editFixture);
 
     expect(response.status).toBe(404);
     expect(response.body).toEqual({ message: 'Task not found.' });
   });
 
   it('should not edit a task owned by another account', async () => {
-    const otherAccount = await registerMockAccount({
+    const otherAccount = await registerAccount({
       email: 'other.edit.tasks@example.com',
     });
 
-    const response = await editTaskRequest(
-      task.id,
-      otherAccount.accessToken,
-    ).send(editFixture);
+    const response = await editTask(task.id)
+      .auth(otherAccount.accessToken)
+      .send(editFixture);
 
     expect(response.status).toBe(404);
     expect(response.body).toEqual({ message: 'Task not found.' });
@@ -115,8 +113,8 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
 
   it('should not edit a task if any fields are empty', async () => {
     const errorResponses = await Promise.all([
-      editTaskRequest(task.id, account.accessToken).send({ name: '' }),
-      editTaskRequest(task.id, account.accessToken).send({ priority: '' }),
+      editTask(task.id).auth(account.accessToken).send({ name: '' }),
+      editTask(task.id).auth(account.accessToken).send({ priority: '' }),
     ]);
 
     errorResponses.forEach((response) => {
@@ -132,17 +130,19 @@ describe('`PUT /tasks/:taskId` endpoint', () => {
   });
 
   it('should not edit a task if priority is unknown', async () => {
-    const response = await editTaskRequest(task.id, account.accessToken).send({
-      ...editFixture,
-      priority: 'some-unknown-priority',
-    });
+    const response = await editTask(task.id)
+      .auth(account.accessToken)
+      .send({
+        ...editFixture,
+        priority: 'some-unknown-priority',
+      });
 
     expect(response.status).toBe(400);
     expect(response.body).toEqual({ message: 'Unknown priority.' });
   });
 
   it('should not edit a task if the user is not authenticated', async () => {
-    const response = await editTaskRequest(task.id).send(editFixture);
+    const response = await editTask(task.id).send(editFixture);
 
     expect(response.status).toBe(401);
     expect(response.body).toEqual({
